+++ spec/fixtures/dummy/app/assets/config/manifest.js
@@ -1,4 +1,3 @@
 //= link_tree ../images
 //= link_directory ../javascripts .js
-//= link_tree ../javascripts/ux_manifests .js
 //= link_directory ../stylesheets .css