+++ spec/fixtures/dummy/app/assets/javascripts/ux/hooks/use_prefetch.js
@@ -1,0 +1,6 @@
+export const usePrefetch = function(url) {
+  return {
+    url: url,
+    started: true
+  };
+};