+++ spec/fixtures/dummy/app/assets/javascripts/ux/lib/pluralize_words.js
@@ -1,0 +1,3 @@
+export const pluralizeWords = function(count, word) {
+  return count === 1 ? word : word + 's';
+};