+++ test/dummy/app/assets/javascripts/ux_dashboard.js
@@ -1,8 +1,0 @@
-// AUTO-GENERATED — DO NOT EDIT
-// react-manifest-rails 0.1.0 | 2026-04-13T05:19:06Z
-// Run `rails react_manifest:generate` to regenerate.
-
-
-//= require ux_shared
-
-//= require ux/app/dashboard/dashboard_index